refactor(onboarding): consolidate tour and progress into a single onboarding widget (AYC-000)
- Merge features/onboarding-tour and features/onboarding-progress into
  widgets/onboarding — features/ is headless in this codebase; the
  provider/target/renderer pattern matches widgets/confirmation-modal
- Fold the single-consumer widgets/onboarding-return-button slice into
  the new widget's ui/
- Move shared/config/tour-targets.ts into the widget's config/ so all
  onboarding consumers import from one public index
- Document the new slice in widgets/SUMMARY.md
