--- ayunis-core-frontend/src/features/onboarding-progress/index.ts
@@ -1,12 +1,0 @@
-export {
-  useOnboardingProgress,
-  type OnboardingProgress,
-} from './lib/useOnboardingProgress';
-export { useOnboarding } from './api/useOnboarding';
-export {
-  ACTION_TYPE,
-  SECONDARY_ACTION_TYPE,
-  type OnboardingStep,
-  type OnboardingCategory,
-  type OnboardingStepId,
-} from './config/categories';
