--- ayunis-core-frontend/src/features/onboarding-tour/index.ts
@@ -1,6 +1,0 @@
-export { OnboardingTourProvider } from './ui/OnboardingTourProvider';
-export {
-  useOnboardingTour,
-  type TourRequest,
-} from './model/onboardingTourContext';
-export { OnboardingTourTarget } from './ui/OnboardingTourTarget';
