+++ ayunis-core-frontend/src/pages/admin-settings/model-settings/ui/ModelSettingsPage.tsx
@@ -6,8 +6,7 @@ import ModelTypeCard from './ModelTypeCard';
 import { OrgDefaultModelCard } from './OrgDefaultModelCard';
 import SettingsLayout from '../../admin-settings-layout';
 import { HelpLink } from '@/shared/ui/help-link/HelpLink';
-import { OnboardingTourTarget } from '@/features/onboarding-tour';
-import { TOUR_TARGET } from '@/shared/config/tour-targets';
+import { OnboardingTourTarget, TOUR_TARGET } from '@/widgets/onboarding';
 import { Trans, useTranslation } from 'react-i18next';
 import {
   useLanguageModels,
