+++ ayunis-core-frontend/src/pages/admin-settings/teams-settings/ui/TeamsSettingsPage.tsx
@@ -1,8 +1,7 @@
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Button } from '@/shared/ui/shadcn/button';
-import { OnboardingTourTarget } from '@/features/onboarding-tour';
-import { TOUR_TARGET } from '@/shared/config/tour-targets';
+import { OnboardingTourTarget, TOUR_TARGET } from '@/widgets/onboarding';
 import { TeamsList } from './TeamsList';
 import TeamsFilters from './TeamsFilters';
 import {
