--- ayunis-core-frontend/src/widgets/onboarding-return-button/index.ts
@@ -1,1 +1,0 @@
-export { OnboardingReturnButton } from './ui/OnboardingReturnButton';
