+++ ayunis-core-frontend/src/widgets/onboarding/config/categories.ts
@@ -1,4 +1,4 @@
-import { TOUR_TARGET, type TourTargetName } from '@/shared/config/tour-targets';
+import { TOUR_TARGET, type TourTargetName } from './tour-targets';
 
 export const ACTION_TYPE = {
   prompt: 'prompt',
