+++ ayunis-core-frontend/src/widgets/onboarding/config/tour-targets.ts
