+++ ayunis-core-frontend/src/widgets/onboarding/index.ts
@@ -1,0 +1,20 @@
+export { OnboardingTourProvider } from './ui/OnboardingTourProvider';
+export {
+  useOnboardingTour,
+  type TourRequest,
+} from './model/onboardingTourContext';
+export { OnboardingTourTarget } from './ui/OnboardingTourTarget';
+export { OnboardingReturnButton } from './ui/OnboardingReturnButton';
+export { useOnboarding } from './api/useOnboarding';
+export {
+  useOnboardingProgress,
+  type OnboardingProgress,
+} from './lib/useOnboardingProgress';
+export {
+  ACTION_TYPE,
+  SECONDARY_ACTION_TYPE,
+  type OnboardingStep,
+  type OnboardingCategory,
+  type OnboardingStepId,
+} from './config/categories';
+export { TOUR_TARGET, type TourTargetName } from './config/tour-targets';
