+++ ayunis-core-frontend/src/widgets/onboarding/lib/useOnboardingProgress.ts
