+++ ayunis-core-frontend/src/widgets/onboarding/lib/useStableTarget.ts
