+++ ayunis-core-frontend/src/widgets/onboarding/ui/TourRenderer.tsx
